+++ spring-application/src/main/kotlin/com/tubefans/gamepicker/config/DatabaseConfig.kt
@@ -1,9 +1,9 @@
 package com.tubefans.gamepicker.config
 
 import com.tubefans.gamepicker.cache.GoogleSheetCache
-import com.tubefans.gamepicker.repositories.GoogleSheetYoCountRepository
 import com.tubefans.gamepicker.repositories.YoCountRepository
 import com.tubefans.gamepicker.services.GoogleSheetsService
+import com.tubefans.gamepicker.services.YoCountService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
@@ -16,5 +16,5 @@ class DatabaseConfig
         val googleSheetsService: GoogleSheetsService,
     ) {
         @Bean
-        fun yoCountRepository(): YoCountRepository = GoogleSheetYoCountRepository(googleSheetCache, googleSheetsService)
+        fun yoCountRepository(): YoCountRepository = YoCountService(googleSheetCache, googleSheetsService)
     }
